modularize add-task function
