+++ src/index.js
@@ -3,7 +3,7 @@ import './style.css';
 
 import displayTasks from './modules/display.js';
 import addTask from './modules/add.js';
-import Task from './modules/task';
+import Task from './modules/task.js';
 
 const taskList = document.querySelector('.task-list');
 
@@ -20,14 +20,6 @@ const taskArr = [
 
 // modify below this line
 
-// class Task {
-//   constructor(index, description) {
-//     this.index = index;
-//     this.description = description;
-//     this.completed = false;
-//   }
-// }
-
 const form = document.querySelector('.form');
 
 form.addEventListener('submit', (e) => {
@@ -36,21 +28,9 @@ form.addEventListener('submit', (e) => {
     new FormData(e.target)
   )
 
-
-  const parsedArr = JSON.parse(localStorage.getItem('taskArr'));
-  const counter = parsedArr.length + 1;
-
-  const newTask = new Task(counter, input.task);
-  parsedArr.push(newTask);
-  localStorage.setItem('taskArr', JSON.stringify(parsedArr));
-
-  const taskfield = document.querySelector('.text-field');
-  taskfield.value = '';
+  addTask(input.task);
   displayTasks();
 });
-
-
-
 
 // modify above this line
 displayTasks();
